Fix for dumping exif data loaded from images that contain a Gps VersionID tag
- This puts in python2 specific conditional check into _value_to_bytes() that
will make sure a raw_value of type str uses struct.unpack to generate int's
to send to the _pack_byte() function, otherwise it crashes trying to create
ints from the chars in the str object.

This crash was found when loading exif data from a picture shot with a
Canon PowerShot SX50 HS (according to the exif data), and it's Gps/VersionID
value was 2.3.0.0 but at run time was loaded as '\x02\x03\x00\x00'.

This probably needs an equivalent fix for Python 3.